+++ src/fides/config/create.py
@@ -250,10 +250,13 @@ def create_config_file(config: FidesConfig, fides_directory_location: str = ".")
 def create_and_update_config_file(
     config: FidesConfig,
     fides_directory_location: str = ".",
-    opt_in: bool = False,
+    opt_in: Optional[bool] = None,
+    opt_out: Optional[bool] = None,
 ) -> Tuple[FidesConfig, str]:
     # request explicit consent for analytics collection
-    config = request_analytics_consent(config=config, opt_in=opt_in)
+    config = request_analytics_consent(
+        config=config, opt_in=opt_in, opt_out=opt_out
+    )
 
     # create the config file as needed
     config_path = create_config_file(